requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
Flask==3.0.3
Flask-Caching==2.3.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.10
python-dotenv==1.2.2
requests==2.32.5
supabase==2.4.5
urllib3==2.6.3
Werkzeug==3.0.1
